Improve code and reduce redundancy

+++ print_chars.c
@@ -1,3 +1,4 @@
+#include <unistd.h>
 #include "main.h"
 
 /**
@@ -9,7 +10,7 @@
  */
 int _putchar(char c)
 {
-	return (write(1, &c, 1));
+	return ((int) (write(1, &c, 1)));
 }
